[LCC/Codegen/MIR] Add `defining_use` to `Register` struct

+++ include/lcc/codegen/isel.hh
@@ -253,7 +253,7 @@ struct PatternList {
                     isz output_i = 0;
                     pattern::output::foreach ([&]<typename inst> {
                         // Use instruction's vreg from input of pattern.
-                        auto output = new MInst(inst::opcode, {input.back()->reg(), input.back()->regsize()});
+                        auto output = new MInst(inst::opcode, {input.back()->reg(), uint(input.back()->regsize())});
                         // Keep track of newly allocated machine instructions.
                         pool.push_back(output);
                         instructions.insert(instructions.begin() + output_i, output);
